+++ src/test/css/navigation.test.ts
@@ -336,7 +336,7 @@ suite('CSS - Navigation', () => {
 		test('No links with empty range', () => {
 			let p = new Parser();
 			assertLinks(p, `body { background-image: url()`, []);
-			assertLinks(p, `@import url();`, [])
+			assertLinks(p, `@import url();`, []);
 		});
 
 	});